+++ crates/lexer/tests/hex.rs
@@ -10,7 +10,7 @@ fn parses_single_hex() {
 
     // The first and only token should be lexed as Literal(0xa57B)
     let tok = lexer.next().unwrap().unwrap();
-    assert_eq!(tok, Token::new(TokenKind::Literal(str_to_bytes32("a57B")), Span::new(2..5, None)));
+    assert_eq!(tok, Token::new(TokenKind::Literal(str_to_bytes32("a57B")), Span::new(0..6, None)));
 
     // We covered the whole source
     lexer.next();
@@ -25,7 +25,7 @@ fn parses_odd_len_hex() {
 
     // The first and only token should be lexed as Literal(0x1)
     let tok = lexer.next().unwrap().unwrap();
-    assert_eq!(tok, Token::new(TokenKind::Literal(str_to_bytes32("1")), Span::new(2..2, None)));
+    assert_eq!(tok, Token::new(TokenKind::Literal(str_to_bytes32("1")), Span::new(0..3, None)));
 
     // We covered the whole source
     lexer.next();