+++ crates/parser/tests/error.rs
@@ -25,16 +25,16 @@ fn test_parses_custom_error() {
                 arg_type: Some(String::from("uint256")),
                 name: None,
                 indexed: false,
-                span: AstSpan(vec![Span { start: 24, end: 30, file: None }]),
+                span: AstSpan(vec![Span { start: 24, end: 31, file: None }]),
                 arg_location: None,
             }],
             span: AstSpan(vec![
-                Span { start: 0, end: 6, file: None },
-                Span { start: 8, end: 12, file: None },
-                Span { start: 14, end: 22, file: None },
-                Span { start: 23, end: 23, file: None },
-                Span { start: 24, end: 30, file: None },
-                Span { start: 31, end: 31, file: None }
+                Span { start: 0, end: 7, file: None },
+                Span { start: 8, end: 13, file: None },
+                Span { start: 14, end: 23, file: None },
+                Span { start: 23, end: 24, file: None },
+                Span { start: 24, end: 31, file: None },
+                Span { start: 31, end: 32, file: None }
             ])
         }
     );
@@ -60,11 +60,11 @@ fn test_error_sel_no_param() {
             selector: [48, 205, 116, 113],
             parameters: vec![],
             span: AstSpan(vec![
-                Span { start: 0, end: 6, file: None },
-                Span { start: 8, end: 12, file: None },
-                Span { start: 14, end: 21, file: None },
-                Span { start: 22, end: 22, file: None },
-                Span { start: 23, end: 23, file: None }
+                Span { start: 0, end: 7, file: None },
+                Span { start: 8, end: 13, file: None },
+                Span { start: 14, end: 22, file: None },
+                Span { start: 22, end: 23, file: None },
+                Span { start: 23, end: 24, file: None }
             ])
         }
     );